docstrings and moved set_model_fit to conftest
