+++ tests/clv/test_plotting.py
@@ -1,17 +1,74 @@
-from typing import Union
-
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import pymc as pm
 import pytest
-import xarray as xr
-from pytensor.tensor import TensorVariable
 
+from pymc_marketing.clv.models import BetaGeoModel, ParetoNBDModel
 from pymc_marketing.clv.plotting import (
     plot_customer_exposure,
     plot_frequency_recency_matrix,
     plot_probability_alive_matrix,
 )
+from tests.clv.conftest import set_model_fit
+
+
+@pytest.fixture(scope="module")
+def fitted_bg(test_summary_data) -> BetaGeoModel:
+    rng = np.random.default_rng(13)
+    data = pd.DataFrame(
+        {
+            "customer_id": test_summary_data.index,
+            "frequency": test_summary_data["frequency"],
+            "recency": test_summary_data["recency"],
+            "T": test_summary_data["T"],
+        }
+    )
+    model_config = {
+        # Narrow Gaussian centered at MLE params from lifetimes BetaGeoFitter
+        "a_prior": {"dist": "DiracDelta", "kwargs": {"c": 1.85034151}},
+        "alpha_prior": {"dist": "DiracDelta", "kwargs": {"c": 1.86428187}},
+        "b_prior": {"dist": "DiracDelta", "kwargs": {"c": 3.18105431}},
+        "r_prior": {"dist": "DiracDelta", "kwargs": {"c": 0.16385072}},
+    }
+    model = BetaGeoModel(
+        data=data,
+        model_config=model_config,
+    )
+    model.build_model()
+    fake_fit = pm.sample_prior_predictive(
+        samples=50, model=model.model, random_seed=rng
+    ).prior
+    set_model_fit(model, fake_fit)
+
+    return model
+
+
+@pytest.fixture(scope="module")
+def fitted_pnbd(test_summary_data) -> ParetoNBDModel:
+    rng = np.random.default_rng(45)
+
+    model_config = {
+        # Narrow Gaussian centered at MLE params from lifetimes ParetoNBDFitter
+        "r_prior": {"dist": "DiracDelta", "kwargs": {"c": 0.5534}},
+        "alpha_prior": {"dist": "DiracDelta", "kwargs": {"c": 10.5802}},
+        "s_prior": {"dist": "DiracDelta", "kwargs": {"c": 0.6061}},
+        "beta_prior": {"dist": "DiracDelta", "kwargs": {"c": 11.6562}},
+    }
+    pnbd_model = ParetoNBDModel(
+        data=test_summary_data,
+        model_config=model_config,
+    )
+    pnbd_model.build_model()
+
+    # Mock an idata object for tests requiring a fitted model
+    # TODO: This is quite slow. Check similar fixtures in the model tests to speed this up.
+    fake_fit = pm.sample_prior_predictive(
+        samples=50, model=pnbd_model.model, random_seed=rng
+    ).prior
+    set_model_fit(pnbd_model, fake_fit)
+
+    return pnbd_model
 
 
 @pytest.mark.parametrize(
@@ -54,50 +111,7 @@ def test_plot_customer_exposure_invalid_args(test_summary_data, kwargs) -> None:
         plot_customer_exposure(test_summary_data, **kwargs)
 
 
-class MockModel:
-    def __init__(self, frequency, recency):
-        self.frequency = frequency
-        self.recency = recency
-
-    def _mock_posterior(
-        self, customer_id: Union[np.ndarray, pd.Series]
-    ) -> xr.DataArray:
-        n_customers = len(customer_id)
-        n_chains = 4
-        n_draws = 10
-        chains = np.arange(n_chains)
-        draws = np.arange(n_draws)
-        return xr.DataArray(
-            data=np.ones((n_customers, n_chains, n_draws)),
-            coords={"customer_id": customer_id, "chain": chains, "draw": draws},
-            dims=["customer_id", "chain", "draw"],
-        )
-
-    def expected_probability_alive(
-        self,
-        customer_id: Union[np.ndarray, pd.Series],
-        frequency: Union[np.ndarray, pd.Series],
-        recency: Union[np.ndarray, pd.Series],
-        T: Union[np.ndarray, pd.Series],
-    ):
-        return self._mock_posterior(customer_id)
-
-    def expected_num_purchases(
-        self,
-        customer_id: Union[np.ndarray, pd.Series],
-        t: Union[np.ndarray, pd.Series, TensorVariable],
-        frequency: Union[np.ndarray, pd.Series, TensorVariable],
-        recency: Union[np.ndarray, pd.Series, TensorVariable],
-        T: Union[np.ndarray, pd.Series, TensorVariable],
-    ):
-        return self._mock_posterior(customer_id)
-
-
-@pytest.fixture
-def mock_model(test_summary_data) -> MockModel:
-    return MockModel(test_summary_data["frequency"], test_summary_data["recency"])
-
-
+@pytest.mark.parametrize("mock_model", [fitted_bg, fitted_pnbd])
 def test_plot_frequency_recency_matrix(mock_model) -> None:
     ax: plt.Axes = plot_frequency_recency_matrix(mock_model)
